FinalProyect/requirements.txt
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.15.30
awscrt==0.19.19
Babel==2.9.1
blinker==1.9.0
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
click==8.1.7
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.8
ffmpeg-python==0.2.0
Flask==3.1.0
future==1.0.0
gpg==1.15.1
idna==2.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==3.0.2
netifaces==0.10.6
numpy==2.0.2
oauthlib==3.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
pytesseract==0.3.13
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
support-info==1.0
systemd-python==235
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==3.1.3
zipp==3.21.0
